rename JU-NEW, fix an Oops in TargetType/ValueType
